requirements.txt
python-telegram-bot==20.6
requests
aiogram==2.25.1
supabase==2.3.0
python-dotenv==1.0.0
